requirements.txt
gradio==4.26.0
gradio_client==0.15.1
primp==0.5.5
pypinyin
tiktoken
socksio
tqdm
colorama
googlesearch-python
Pygments
openai==1.16.2
langchain==0.1.14
langchain-openai
langchainhub
langchain_community
groq
markdown
PyPDF2
pdfplumber
pandas
commentjson
openpyxl
pandoc
wolframalpha
faiss-cpu==1.7.4
duckduckgo-search>=5.3.0
arxiv
wikipedia
google-cloud-aiplatform
google.generativeai
unstructured
google-api-python-client
tabulate
ujson
python-docx
websocket_client
pydantic==2.5.2
google-search-results
anthropic==0.18.1
Pillow>=10.1.0
protobuf==3.20.3
ollama>=0.1.6
numexpr
regex
